Refactor S2ST (#4747)
* Move S2ST submodules to separate files

* Refactoring

* Fix import
