+++ fairseq/models/speech_to_speech/s2s_conformer.py
@@ -10,14 +10,30 @@ import torch
 
 from fairseq import checkpoint_utils
 from fairseq.models import register_model, register_model_architecture
-from fairseq.models.speech_to_speech.s2s_transformer import S2UTTransformerModel
 from fairseq.models.speech_to_speech.s2s_transformer import (
-    s2ut_architecture_base as s2ut_transformer_architecture_base,
+    S2UTTransformerModel,
+    s2ut_architecture_base,
 )
 from fairseq.models.speech_to_text import S2TConformerEncoder
 from fairseq.models.transformer import Linear
 
 logger = logging.getLogger(__name__)
+
+
+def build_s2s_conformer_encoder(args):
+    encoder = S2SConformerEncoder(args)
+    pretraining_path = getattr(args, "load_pretrained_encoder_from", None)
+    if pretraining_path is not None:
+        if not Path(pretraining_path).exists():
+            logger.warning(
+                f"skipped pretraining because {pretraining_path} does not exist"
+            )
+        else:
+            encoder = checkpoint_utils.load_pretrained_component_from_model(
+                component=encoder, checkpoint=pretraining_path
+            )
+            logger.info(f"loaded pretrained encoder from: {pretraining_path}")
+    return encoder
 
 
 class S2SConformerEncoder(S2TConformerEncoder):
@@ -51,7 +67,7 @@ class S2SConformerEncoder(S2TConformerEncoder):
 @register_model("s2ut_conformer")
 class S2UTConformerModel(S2UTTransformerModel):
     """
-    Direct speech-to-speech translation model with S2T Conformer encoder + Transformer discrete unit decoder
+    Direct speech-to-speech translation model with Conformer encoder + Transformer discrete unit decoder
     """
 
     @staticmethod
@@ -78,25 +94,15 @@ class S2UTConformerModel(S2UTTransformerModel):
 
     @classmethod
     def build_encoder(cls, args):
-        encoder = S2SConformerEncoder(args)
-        pretraining_path = getattr(args, "load_pretrained_encoder_from", None)
-        if pretraining_path is not None:
-            if not Path(pretraining_path).exists():
-                logger.warning(
-                    f"skipped pretraining because {pretraining_path} does not exist"
-                )
-            else:
-                encoder = checkpoint_utils.load_pretrained_component_from_model(
-                    component=encoder, checkpoint=pretraining_path
-                )
-                logger.info(f"loaded pretrained encoder from: {pretraining_path}")
-        return encoder
+        return build_s2s_conformer_encoder(args)
 
 
 @register_model_architecture("s2ut_conformer", "s2ut_conformer")
-def s2ut_base_architecture(args):
+def s2ut_conformer_architecture_base(args):
     args.attn_type = getattr(args, "attn_type", None)
     args.pos_enc_type = getattr(args, "pos_enc_type", "abs")
+    args.input_feat_per_channel = getattr(args, "input_feat_per_channel", 80)
+    args.input_channels = getattr(args, "input_channels", 1)
     args.max_source_positions = getattr(args, "max_source_positions", 6000)
     args.encoder_embed_dim = getattr(args, "encoder_embed_dim", 256)
     args.encoder_ffn_embed_dim = getattr(args, "encoder_ffn_embed_dim", 2048)
@@ -104,4 +110,4 @@ def s2ut_base_architecture(args):
     args.dropout = getattr(args, "dropout", 0.1)
     args.encoder_layers = getattr(args, "encoder_layers", 16)
     args.depthwise_conv_kernel_size = getattr(args, "depthwise_conv_kernel_size", 31)
-    s2ut_transformer_architecture_base(args)
+    s2ut_architecture_base(args)
